requirements.txt
# ProteinPRO: Protein Polymer Reactivity Optimization

# Data Science and Machine Learning
numpy>=1.24.0
openpyxl>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.0
shap>=0.43.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Structural Biology and Protein Analysis
biopython>=1.81
MDAnalysis>=2.6.0

# Chemical Descriptor Featurization
rdkit>=2023.9.1
mordred>=1.2.0

# 3D Visualization: Web App
py3Dmol>=2.0.0
nglview>=3.0.0
streamlit>=1.28.0
streamlit[auth]>=1.28.0

# API and Optional Integrations
requests>=2.31.0

# Hack Duke 2026 Tech Stack (optional)
google-generativeai>=0.8.0   # Gemini API - Best Use of Generative AI

# Utilities and Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0
